Hub & Daemon: API reentrance (on server side).
Considering the following scenario:
* workflow is: op0 -> op1 -> op2
* op0 and op2 are on daemonA, op1 on daemonB

In order to get the output of op2:
* the hub sends a request to api of daemonA
* daemonA pulls op1's output from the hub
* the hub sends a request to api of daemonB
* daemonB pulls op0's output from the hub
* the hub sends a request to api of daemonA

These 5 api calls are imbricated: when the 5th
returns, it unblocks the 4th, and so on.

The issue here is that the 1st and 5th calls
are issued on the same API / socket, thus
API reentrance has to be correctly handled.

We will consider:
1. when a daemon has to pull data from the hub,
   it will open a new rpc connection. This will
   cause the hub to open a new greenlet, thus
   it will not block.
2. the daemon will open a new greenlet each time
   it receives an API call.

This commit implements case 2.